+++ odoo/local-src/sf_mrp/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 # Part of sensefly.
 import models
+import wizard